requirements.txt
# Core scientific computing
numpy>=1.24.0
scipy>=1.11.0
pandas>=2.0.0

# Audio processing
librosa>=0.10.1
soundfile>=0.12.1
torchaudio>=2.1.0

# Deep learning
torch>=2.1.0
torchvision>=0.16.0
timm>=0.9.12

# Interpretability
grad-cam>=1.4.8

# Machine learning baselines
scikit-learn>=1.3.0
xgboost>=2.0.0

# Experiment tracking
wandb>=0.16.0
tensorboard>=2.14.0

# Visualisation
matplotlib>=3.7.0
seaborn>=0.13.0
pillow>=10.0.0

# Configuration
PyYAML>=6.0.1
omegaconf>=2.3.0

# Utilities
tqdm>=4.66.0
click>=8.1.7
rich>=13.6.0
joblib>=1.3.0

# Statistical analysis
pingouin>=0.5.3

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
